refactor(StructureRendering): removed StructureRendering.tsx by moving everything into components
