+++ packages/react/src/app/util.ts
@@ -8,6 +8,7 @@ import {
   ThingTypeName,
 } from "../model/ModelTypeCheck";
 import { StoryContext } from "@storybook/react";
+import Reified from "../model/Reified";
 
 export const OBO_FOUNDRY_REPO_URL_RAW =
   "https://raw.githubusercontent.com/OBOFoundry/OBOFoundry.github.io/master" as const;
@@ -19,6 +20,15 @@ export function asArray<T>(obj: T | T[]): T[] {
     return [obj];
   }
   return [] as T[];
+}
+
+export function asReified<T>(obj: T | Reified<T>): Reified<T> {
+    if (obj instanceof Reified) {
+        return obj;
+    }
+    else {
+        return Reified.fromJson<T>(obj)[0];
+    }
 }
 
 const DEFAULT_USE_LEGACY = true;
