+++ packages/react/src/model/RenderedReified.tsx
@@ -4,8 +4,9 @@ import {OnNavigates} from "../app";
 import React, {ReactElement} from "react";
 import EntityLink from "./EntityLink";
 import ClassExpression from "./ClassExpression";
-import {addLinksToText, getAxiomsInformationJSX} from "./StructureRendering";
 import {DEFAULT_SHOW_BADGES} from "../app/globals";
+import Tooltip from "./Tooltip";
+import {randomString} from "../app/util";
 
 /**
  * Renders a given Reified
@@ -40,16 +41,105 @@ export default function RenderedReified(
     ): ReactElement {
         const linkedEntities = parentEntity.getLinkedEntities();
 
+        /**
+         * Inserts links into text (potentially with label instead of link displayed if link exists as key inside linkedEntities) and returns the resulting JSX element
+         * @param text the text to insert links into
+         */
+        function addLinksToText(
+            text: string
+        ) {
+            const linksToSplice: Array<{
+                start: number;
+                end: number;
+                link: ReactElement;
+            }> = [];
+
+            if (linkedEntities) {
+                for (const entityId of Object.keys(linkedEntities.linkedEntities)) {
+                    for (
+                        let n = text.indexOf(entityId, 0);
+                        n !== -1;
+                        n = text.indexOf(entityId, n)
+                    ) {
+                        linksToSplice.push({
+                            start: n,
+                            end: n + entityId.length,
+                            link: <EntityLink
+                                parentEntity={parentEntity}
+                                linkedEntities={linkedEntities}
+                                iri={entityId}
+                                showBadges={showBadges}
+                                onNavigates={onNavigates}
+                            />,
+                        });
+
+                        n += entityId.length;
+                    }
+                }
+            }
+
+            const urlRe = /[A-z]+:\/\/[^\s]+/g;
+            for (let match = urlRe.exec(text); match; match = urlRe.exec(text)) {
+                const url = match[0];
+                // console.log("found match " + url);
+
+                linksToSplice.push({
+                    start: match.index,
+                    end: match.index + url.length,
+                    link: (
+                        <span key={randomString()}>
+                          <a className="clickable" href={url}>
+                            {url}
+                          </a>
+                        </span>
+                    ),
+                });
+            }
+
+            linksToSplice.sort((a, b) => a.start - b.start);
+
+            removeOverlapping: for (let n = 0; n < linksToSplice.length; ) {
+                for (let n2 = n + 1; n2 < linksToSplice.length; ++n2) {
+                    const spliceA = linksToSplice[n];
+                    const spliceB = linksToSplice[n2];
+
+                    if (spliceA === spliceB) continue;
+
+                    // The splices overlap if neither ends before the other starts
+                    if (spliceA.end >= spliceB.start && spliceB.end >= spliceA.start) {
+                        // console.log("Removing overlapping");
+
+                        // remove the shorter link of both
+                        if (spliceA.end - spliceA.start < spliceB.end - spliceB.start) {
+                            linksToSplice.splice(n, 1);
+                        } else {
+                            linksToSplice.splice(n2, 1);
+                        }
+                        continue removeOverlapping;
+                    }
+                }
+                ++n;
+            }
+
+            if (linksToSplice.length === 0) return <>{text}</>;
+
+            const res: ReactElement[] = [];
+            let n = 0;
+
+            for (const link of linksToSplice) {
+                res.push(<span key={randomString()}>{text.substring(n, link.start)}</span>);
+                res.push(link.link);
+                n = link.end;
+            }
+            res.push(<span key={randomString()}>{text.slice(n)}</span>);
+
+            return <>{res}</>;
+        }
+
         // linkedEntities not existent on entity (-> probably legacy api version)
         if (Object.keys(linkedEntities.linkedEntities).length == 0) {
             if (typeof value == "string") {
-                return addLinksToText(
-                    parentEntity,
-                    value.toString(),
-                    undefined,
-                    showBadges,
-                    onNavigates,
-                );
+                return addLinksToText(value.toString());
             } else if (
                 typeof value == "object" &&
                 value["http://www.w3.org/1999/02/22-rdf-syntax-ns#type"] !=
@@ -105,24 +195,39 @@ export default function RenderedReified(
                         />
                     }
                 } else {
-                    return addLinksToText(
-                        parentEntity,
-                        value.toString(),
-                        linkedEntities,
-                        showBadges,
-                        onNavigates,
-                    );
+                    return addLinksToText(value.toString(),);
                 }
             }
         }
+    }
+
+    /**
+     * @param axiomsDict the entities axioms in the format returned by Reified::getMetadata()
+     */
+    function RenderedAxioms(
+        {
+            axiomsDict
+        } : {
+            axiomsDict: any | null
+        }
+    ): ReactElement {
+        const axiomsText = Object.keys(axiomsDict)
+            .map((key) => {
+                const label = parentEntity.getLinkedEntities().getLabelForIri(key) || key;
+                if (label) {
+                    return "*" + axiomsDict[key] + " (" + label + ")";
+                } else return "";
+            })
+            .join("\n");
+
+        return <Tooltip text={axiomsText} />;
     }
 
     return (
         <>
             <RenderedValue value={reified.value} />
             &nbsp;
-            {reified.hasMetadata() &&
-                getAxiomsInformationJSX(parentEntity, reified.getMetadata())}
+            {reified.hasMetadata() && <RenderedAxioms axiomsDict={reified.getMetadata()}/>}
         </>
     );
 }
